update to 2019.2.11f1
update to 2019.2.11f1

+++ Dice Incremental/Assets/Scripts/Dice/Dice.cs
@@ -36,7 +36,7 @@ public class Dice : MonoBehaviour
 	public void AddPower() { m_stats.AddSide(); }
 
 	public int GetSides() { return m_stats.GetSides(); }
-	public int GetPower() { return m_stats.GetPower(); }
+	public float GetPower() { return m_stats.GetPower(); }
 	public int GetGoal() { return m_stats.GetGoal(); }
 
 	private void Awake()
